chore: add assignment page
