+++ public/index.html
@@ -33,7 +33,7 @@
       rel="stylesheet"
     />
   </head>
-  <body>
+  <body class="dark">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
