+++ src/assets/data/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_FILE_SIZE_IN_MB = 20
