--- src/pages/Dashboard/DashboardPages/MyAssignment.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const MyAssignment = (): JSX.Element => {
-  return <div>Assignment Page</div>
-}
-
-export default MyAssignment
